Cast Sig as Int

+++ CommandLine.c
@@ -420,7 +420,7 @@ int CommandLine_run(int argc, char** argv) {
    CRT_done();
 
    if (terminate_requested) {
-      fprintf(stderr, "\nInterrupted by signal %ld, terminating.\n", (long)terminate_signal);
+      fprintf(stderr, "\nInterrupted by signal %d, terminating.\n", (int)terminate_signal);
       exit(128 + (int)terminate_signal);
    }
